Add unit test for LIKE operator and simplify its implementation.
